# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * New Testament Priests Church API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface LoginRequest {
  username: string;
  password: string;
}

export type LoginResponseRole =
  (typeof LoginResponseRole)[keyof typeof LoginResponseRole];

export const LoginResponseRole = {
  admin: "admin",
  members: "members",
} as const;

export interface LoginResponse {
  role: LoginResponseRole;
  token: string;
}

export interface SupportSubmission {
  id: number;
  firstName: string;
  middleName: string;
  lastName: string;
  phone: string;
  email: string;
  address: string;
  occupation: string;
  amountPerMonth: number;
  amountPerYear: number;
  reason: string;
  createdAt: string;
}

export interface CreateSupportSubmission {
  firstName: string;
  middleName: string;
  lastName: string;
  phone: string;
  email: string;
  address: string;
  occupation: string;
  amountPerMonth: number;
  amountPerYear: number;
  reason: string;
}

export interface MembershipRequest {
  id: number;
  firstName: string;
  middleName: string;
  lastName: string;
  phone: string;
  email: string;
  address: string;
  occupation: string;
  previousChurch: string;
  servingAs: string;
  baptized: boolean;
  createdAt: string;
}

export interface CreateMembershipRequest {
  firstName: string;
  middleName: string;
  lastName: string;
  phone: string;
  email: string;
  address: string;
  occupation: string;
  previousChurch: string;
  servingAs: string;
  baptized: boolean;
}

export interface Post {
  id: number;
  title: string;
  highlights: string;
  photoUrl?: string;
  facebookUrl?: string;
  youtubeUrl?: string;
  createdAt: string;
}

export interface CreatePost {
  title: string;
  highlights: string;
  photoUrl?: string;
  facebookUrl?: string;
  youtubeUrl?: string;
}

export interface Event {
  id: number;
  title: string;
  description: string;
  date: string;
  imageUrl?: string;
  createdAt: string;
}

export interface CreateEvent {
  title: string;
  description: string;
  date: string;
  imageUrl?: string;
}

export interface Scripture {
  id: number;
  verse: string;
  reference: string;
  createdAt: string;
}

export interface CreateScripture {
  verse: string;
  reference: string;
}

export interface Department {
  id: number;
  name: string;
  description: string;
  groupPhotoUrl?: string;
  members?: string[];
  activities?: string;
  meetingTime?: string;
}

export interface EyEvent {
  id: number;
  title: string;
  description: string;
  date: string;
  imageUrl?: string;
  type: string;
  year: number;
  createdAt: string;
}

export interface CreateEyEvent {
  title: string;
  description: string;
  date: string;
  imageUrl?: string;
  type: string;
  year: number;
}

export interface EyRound {
  id: number;
  eventId: number;
  roundNumber: number;
  capacity: number;
  createdAt: string;
}

export interface CreateEyRound {
  eventId: number;
  roundNumber: number;
  capacity: number;
}

export interface EyParticipant {
  id: number;
  faydaId: string;
  firstName: string;
  middleName?: string;
  lastName: string;
  city: string;
  phoneNumber: string;
  email: string;
  faydaVerified: boolean;
  eventId: number;
  roundId: number;
  paymentStatus: boolean;
  registrationNumber: string;
  coordinatorId: number;
  createdAt: string;
}

export type CreateEyParticipantPaymentMethod =
  (typeof CreateEyParticipantPaymentMethod)[keyof typeof CreateEyParticipantPaymentMethod];

export const CreateEyParticipantPaymentMethod = {
  bank: "bank",
  telebirr: "telebirr",
} as const;

export interface CreateEyParticipant {
  faydaId: string;
  firstName: string;
  middleName?: string;
  lastName: string;
  city: string;
  phoneNumber: string;
  email: string;
  eventId: number;
  roundId: number;
  paymentMethod: CreateEyParticipantPaymentMethod;
  bankReference?: string;
  paymentSlip?: string;
}

export interface UpdateEyParticipant {
  faydaId?: string;
  firstName?: string;
  middleName?: string;
  lastName?: string;
  city?: string;
  phoneNumber?: string;
  email?: string;
  faydaVerified?: boolean;
  eventId?: number;
  roundId?: number;
  paymentStatus?: boolean;
  registrationNumber?: string;
  coordinatorId?: number;
}

export interface EyCoordinator {
  id: number;
  name: string;
  phone: string;
  email: string;
  createdAt: string;
}

export interface CreateEyCoordinator {
  name: string;
  phone: string;
  email: string;
}