# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: New Testament Priests Church API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /ey-rounds:
    get:
      operationId: listEyRounds
      tags: [ey-rounds]
      summary: List all Excellent Youth rounds
      responses:
        "200":
          description: List of EY rounds
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EyRound"
    post:
      operationId: createEyRound
      tags: [ey-rounds]
      summary: Create an Excellent Youth round (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEyRound"
      responses:
        "201":
          description: Round created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EyRound"
  /ey-rounds/{id}:
    put:
      operationId: updateEyRound
      tags: [ey-rounds]
      summary: Update an EY round (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEyRound"
      responses:
        "200":
          description: Round updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EyRound"
    delete:
      operationId: deleteEyRound
      tags: [ey-rounds]
      summary: Delete an EY round (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /ey-participants:
    get:
      operationId: listEyParticipants
      tags: [ey-participants]
      summary: List all Excellent Youth participants
      responses:
        "200":
          description: List of EY participants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EyParticipant"
    post:
      operationId: createEyParticipant
      tags: [ey-participants]
      summary: Register an Excellent Youth participant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEyParticipant"
      responses:
        "201":
          description: Participant registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EyParticipant"
  /ey-participants/{id}:
    put:
      operationId: updateEyParticipant
      tags: [ey-participants]
      summary: Update an EY participant (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEyParticipant"
      responses:
        "200":
          description: Participant updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EyParticipant"
    delete:
      operationId: deleteEyParticipant
      tags: [ey-participants]
      summary: Delete an EY participant (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    EyEvent:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        date:
          type: string
        imageUrl:
          type: string
        type:
          type: string
        year:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - title
        - description
        - date
        - type
        - year
        - createdAt
...
    EyRound:
      type: object
      properties:
        id:
          type: integer
        eventId:
          type: integer
        roundNumber:
          type: integer
        capacity:
          type: integer
        fromDate:
          type: string
        toDate:
          type: string
        createdAt:
          type: string
      required:
        - id
        - eventId
        - roundNumber
        - capacity
        - createdAt
    CreateEyRound:
      type: object
      properties:
        eventId:
          type: integer
        roundNumber:
          type: integer
        capacity:
          type: integer
        fromDate:
          type: string
        toDate:
          type: string
      required:
        - eventId
        - roundNumber
        - capacity
        - fromDate
        - toDate
    EyParticipant:
      type: object
      properties:
        id:
          type: integer
        faydaId:
          type: string
        firstName:
          type: string
        middleName:
          type: string
        lastName:
          type: string
        city:
          type: string
        phoneNumber:
          type: string
        email:
          type: string
        faydaVerified:
          type: boolean
        eventId:
          type: integer
        roundId:
          type: integer
        paymentStatus:
          type: boolean
        registrationNumber:
          type: string
        coordinatorId:
          type: integer
        badgeGenerated:
          type: boolean
        event:
          $ref: "#/components/schemas/EyEvent"
        round:
          $ref: "#/components/schemas/EyRound"
        coordinator:
          $ref: "#/components/schemas/EyCoordinator"
        createdAt:
          type: string
      required:
        - id
        - faydaId
        - firstName
        - lastName
        - city
        - phoneNumber
        - email
        - faydaVerified
        - eventId
        - roundId
        - paymentStatus
        - registrationNumber
        - createdAt
    CreateEyParticipant:
      type: object
      properties:
        faydaId:
          type: string
        firstName:
          type: string
        middleName:
          type: string
        lastName:
          type: string
        city:
          type: string
        phoneNumber:
          type: string
        email:
          type: string
        eventId:
          type: integer
        roundId:
          type: integer
        paymentMethod:
          type: string
          enum: [bank, telebirr]
        bankReference:
          type: string
        paymentSlip:
          type: string
      required:
        - faydaId
        - firstName
        - lastName
        - city
        - phoneNumber
        - email
        - eventId
        - roundId
        - paymentMethod
    UpdateEyParticipant:
      type: object
      properties:
        faydaId:
          type: string
        firstName:
          type: string
        middleName:
          type: string
        lastName:
          type: string
        city:
          type: string
        phoneNumber:
          type: string
        email:
          type: string
        faydaVerified:
          type: boolean
        eventId:
          type: integer
        roundId:
          type: integer
        paymentStatus:
          type: boolean
        registrationNumber:
          type: string
        coordinatorId:
          type: integer
    EyCoordinator:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        createdAt:
          type: string
      required:
        - id
        - name
        - phone
        - email
        - createdAt
...