# lib/api-zod/src/generated/types/createEyRound.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * New Testament Priests Church API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateEyRound {
  eventId: number;
  roundNumber: number;
  capacity: number;
}