# lib/api-zod/src/generated/types/eyEvent.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * New Testament Priests Church API
 * OpenAPI spec version: 0.1.0
 */

export interface EyEvent {
  id: number;
  title: string;
  description: string;
  date: string;
  imageUrl?: string;
  type: string;
  year: number;
  createdAt: string;
}