# lib/api-zod/src/generated/types/eyParticipant.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * New Testament Priests Church API
 * OpenAPI spec version: 0.1.0
 */

export interface EyParticipant {
  id: number;
  faydaId: string;
  firstName: string;
  middleName?: string;
  lastName: string;
  city: string;
  phoneNumber: string;
  email: string;
  faydaVerified: boolean;
  eventId: number;
  roundId: number;
  paymentStatus: boolean;
  registrationNumber: string;
  coordinatorId: number;
  createdAt: string;
}